+++ docs/source/changelog.rst
@@ -9,6 +9,7 @@ Change Log
 - added connection troubleshooting checklist to tips & tricks chapter
 - fix sphinx config issues when building the docs
 - some raised exceptions had a __cause__ added on Python 2.x as well, this has been corrected (it could cause unwanted serialization errors)
+- added http example that shows simple use of the http gateway
 
 
 **Pyro 4.32**